7 commit and submit

+++ HashMap.hpp
@@ -96,7 +96,7 @@ class HashMap {
    * */
   void load_data (std::vector<KeyT> keys, std::vector<ValueT> values)
   {
-    for (int i = 0; i < keys.size (); ++i)
+    for (size_t i = 0; i < keys.size (); ++i)
       {
         if (!contains_key (keys[i]))
           {
